# internal/server/static/openapi/local-api.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CSGHub Lite Local API",
    "version": "0.1.0",
    "description": "Structured documentation for the local `/api/*` and `/v1/*` endpoints served by `csghub-lite`.\n\nMaintenance note: whenever you add, remove, or rename an `/api/*` or `/v1/*` route in `internal/server/routes.go`, update this file and run the OpenAPI sync test before merging."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current csghub-lite server"
    },
    {
      "url": "http://localhost:11435",
      "description": "Default local server"
    }
  ],
...
  "paths": {
...
    "/api/models/upload": {
      "post": {
        "tags": [
          "Models"
        ],
        "summary": "Upload a local model",
        "description": "Imports a local model from multipart form data. Use `mode=archive` for a single `.zip`, `.tar`, `.tar.gz`, or `.tgz` file; use `mode=files` or `mode=directory` with one or more files and matching `paths` values.",
        "operationId": "uploadModel",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "files"
                ],
                "properties": {
                  "model": {
                    "type": "string",
                    "description": "Optional target model ID in `namespace/name` form. Defaults to `local/<filename>` when omitted."
                  },
                  "mode": {
                    "type": "string",
                    "enum": [
                      "archive",
                      "directory",
                      "files"
                    ],
                    "default": "files"
                  },
                  "overwrite": {
                    "type": "boolean",
                    "default": false
                  },
                  "paths": {
                    "type": "array",
                    "description": "Relative paths corresponding to uploaded files. Folder uploads should provide browser-relative paths.",
                    "items": {
                      "type": "string"
                    }
                  },
                  "files": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The imported local model.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelUploadResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
    "/api/load": {
      "post": {
        "tags": [
          "Models"
        ],
        "summary": "Eagerly load a model",
        "description": "Loads a model into memory and optionally streams progress while conversions are running.",
        "operationId": "loadModel",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoadRequest"
              },
              "example": {
                "model": "Qwen/Qwen3-0.6B-GGUF",
                "stream": false,
                "num_ctx": 131072,
                "num_parallel": 1,
                "n_gpu_layers": 40,
                "cache_type_k": "q8_0",
                "cache_type_v": "q8_0",
                "dtype": "q8_0"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Load completed or streaming progress returned.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoadResponse"
                },
                "example": {
                  "status": "ready"
                }
              },
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/EventStream"
                },
                "example": "data: {\"status\":\"loading Qwen/Qwen3-0.6B-GGUF\"}\n\ndata: {\"status\":\"converting\",\"step\":\"quantizing\",\"current\":1,\"total\":3}\n\ndata: {\"status\":\"ready\"}\n\n"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
...
    "/api/marketplace/models": {
      "get": {
        "tags": [
          "Marketplace"
        ],
        "summary": "List marketplace models",
        "description": "Proxies the upstream CSGHub model listing API and normalizes the response to `{ data, total }`.",
        "operationId": "getMarketplaceModels",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Marketplace sort mode such as `trending` or `recently_updated`.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "per",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 16
            }
          },
          {
            "name": "source",
            "in": "query",
            "description": "Optional upstream source filter forwarded to CSGHub.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "framework",
            "in": "query",
            "description": "Optional model type filter forwarded to CSGHub, such as `gguf` or `safetensors`.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Marketplace model list.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketplaceModelsResponse"
                }
              }
            }
          },
          "502": {
            "$ref": "#/components/responses/BadGateway"
          }
        }
      }
    },
    "/api/marketplace/models/{namespace}/{name}": {
      "get": {
        "tags": [
          "Marketplace"
        ],
        "summary": "Get marketplace model details",
        "description": "Fetches a single marketplace model and, for GGUF repositories, summarizes the available quantization variants discovered in the upstream file tree.",
        "operationId": "getMarketplaceModelDetail",
        "parameters": [
          {
            "name": "namespace",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Marketplace model details and GGUF quantization summary.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketplaceModelDetailResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "502": {
            "$ref": "#/components/responses/BadGateway"
          }
        }
      }
    },
...
    "/api/apps/model": {
      "post": {
        "tags": [
          "AI Apps"
        ],
        "summary": "Save preferred model for an AI App",
        "description": "Persists the user's preferred model choice for a given AI App so it is used as the default on next launch.",
        "operationId": "saveAIAppModel",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AIAppActionRequest"
              },
              "example": {
                "app_id": "openclaw",
                "model_id": "Qwen/Qwen3-0.6B-GGUF"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Model preference saved."
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
...
    "/v1/models": {
      "get": {
        "tags": [
          "Compatibility"
        ],
        "summary": "List compatibility models",
        "description": "Returns an OpenAI model list by default. When Anthropic-style headers such as `anthropic-version` are present, the same route returns an Anthropic-compatible model list instead.",
        "operationId": "listCompatibilityModels",
        "parameters": [
          {
            "name": "refresh",
            "in": "query",
            "description": "Force a refresh of the cloud model catalog before returning results.",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "anthropic-version",
            "in": "header",
            "description": "Send this header to receive Anthropic-compatible model metadata.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "anthropic-beta",
            "in": "header",
            "description": "Additional Anthropic compatibility hint that also switches the response shape.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "description": "Anthropic-style API key header. When provided without `authorization`, the route returns the Anthropic response format.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OpenAI- or Anthropic-compatible model listing.",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/OpenAIModelList"
                    },
                    {
                      "$ref": "#/components/schemas/AnthropicModelListResponse"
                    }
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Compatibility model listing error.",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/OpenAIErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/AnthropicErrorResponse"
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "responses": {
      "BadRequest": {
        "description": "The request body, query parameters, or path parameters were invalid.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
...
      "InternalError": {
        "description": "An internal server error occurred.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
      "BadGateway": {
        "description": "The upstream CSGHub service returned an error.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
...
    },
    "schemas": {
      "ApiError": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "error",
          "errorCode"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "errorCode": {
            "type": "integer",
            "format": "int32",
            "example": 400
          }
        }
      },
      "EventStream": {
        "type": "string",
        "description": "Server-Sent Event stream payload."
      },
...
      "LoadRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "model"
        ],
        "properties": {
          "model": {
            "type": "string"
          },
          "stream": {
            "type": "boolean",
            "default": false
          },
          "keep_alive": {
            "type": "string"
          },
          "num_ctx": {
            "type": "integer",
            "minimum": 1
          },
          "num_parallel": {
            "type": "integer",
            "minimum": 1
          },
          "n_gpu_layers": {
            "type": "integer",
            "minimum": 0
          },
          "cache_type_k": {
            "type": "string"
          },
          "cache_type_v": {
            "type": "string"
          },
          "dtype": {
            "type": "string"
          }
        }
      },
      "LoadResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "type": "string"
          },
          "step": {
            "type": "string"
          },
          "current": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          }
        }
      },
      "ModelInfo": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "name",
          "model",
          "size",
          "format",
          "modified_at"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "size": {
            "type": "integer"
          },
          "format": {
            "type": "string"
          },
          "modified_at": {
            "type": "string",
            "format": "date-time"
          },
          "display_name": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "source": {
            "type": "string"
          },
          "provider": {
            "type": "string",
            "description": "Normalized provider name for filtering `/api/tags`, such as `local`, `csghub`, or a third-party provider name."
          },
          "category": {
            "type": "string",
            "description": "csghub-lite category derived from the model pipeline tag."
          },
          "pipeline_tag": {
            "type": "string"
          },
          "input_modalities": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Input modalities supported by the model, such as text, image, video, audio, or file."
          },
          "output_modalities": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Output modalities produced by the model, such as text, image, video, audio, speech, or transcription."
          },
          "has_mmproj": {
            "type": "boolean"
          },
          "context_window": {
            "type": "integer"
          },
          "description": {
            "type": "string"
          },
          "license": {
            "type": "string"
          },
          "llm_type": {
            "type": "string"
          },
          "owned_by": {
            "type": "string"
          },
          "pricing": {
            "$ref": "#/components/schemas/ModelPricing"
          }
        }
      },
      "ModelPricing": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "input_token_price": {
            "$ref": "#/components/schemas/ModelTokenPrice"
          },
          "output_token_price": {
            "$ref": "#/components/schemas/ModelTokenPrice"
          }
        }
      },
      "ModelTokenPrice": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "price_per_million"
        ],
        "properties": {
          "currency": {
            "type": "string"
          },
          "price_per_million": {
            "type": "number"
          }
        }
      },
...
      "ModelFileEntry": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "path",
          "size",
          "download_url"
        ],
        "properties": {
          "path": {
            "type": "string"
          },
          "size": {
            "type": "integer"
          },
          "sha256": {
            "type": "string"
          },
          "lfs": {
            "type": "boolean"
          },
          "download_url": {
            "type": "string"
          }
        }
      },
...
      "ModelUploadResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "status",
          "model",
          "details",
          "files"
        ],
        "properties": {
          "status": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "details": {
            "$ref": "#/components/schemas/ModelInfo"
          },
          "files": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ModelFileEntry"
            }
          }
        }
      },
...
      "AIAppActionRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "app_id"
        ],
        "properties": {
          "app_id": {
            "type": "string"
          },
          "model_id": {
            "type": "string"
          },
          "source": {
            "type": "string",
            "description": "Model source such as local, cloud, or provider:<id>."
          },
          "work_dir": {
            "type": "string",
            "description": "Working directory used for shell-based AI App launches."
          }
        }
      },
...
      "MarketplaceTag": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "name",
          "category",
          "show_name"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "group": {
            "type": "string"
          },
          "built_in": {
            "type": "boolean"
          },
          "show_name": {
            "type": "string"
          }
        }
      },
      "MarketplaceRepository": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "http_clone_url": {
            "type": "string"
          },
          "ssh_clone_url": {
            "type": "string"
          }
        }
      },
      "MarketplaceModelMetadata": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "model_params": {
            "type": "number"
          },
          "tensor_type": {
            "type": "string"
          },
          "architecture": {
            "type": "string"
          },
          "mini_gpu_memory_gb": {
            "type": "number"
          },
          "mini_gpu_finetune_gb": {
            "type": "number"
          },
          "model_type": {
            "type": "string"
          },
          "class_name": {
            "type": "string"
          }
        }
      },
      "MarketplaceModel": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "name",
          "description",
          "likes",
          "downloads",
          "path",
          "created_at",
          "updated_at",
          "license"
        ],
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "nickname": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "likes": {
            "type": "integer"
          },
          "downloads": {
            "type": "integer"
          },
          "path": {
            "type": "string"
          },
          "repository_id": {
            "type": "integer"
          },
          "private": {
            "type": "boolean"
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MarketplaceTag"
            }
          },
          "repository": {
            "$ref": "#/components/schemas/MarketplaceRepository"
          },
          "default_branch": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "license": {
            "type": "string"
          },
          "source": {
            "type": "string"
          },
          "sync_status": {
            "type": "string"
          },
          "metadata": {
            "$ref": "#/components/schemas/MarketplaceModelMetadata"
          },
          "hf_path": {
            "type": "string"
          }
        }
      },
...
      "MarketplaceModelsResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "data",
          "total"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MarketplaceModel"
            }
          },
          "total": {
            "type": "integer"
          }
        }
      },
      "MarketplaceModelQuantization": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "name",
          "file_count",
          "example_path"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "file_count": {
            "type": "integer"
          },
          "example_path": {
            "type": "string"
          }
        }
      },
      "MarketplaceModelDetailResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "details"
        ],
        "properties": {
          "details": {
            "$ref": "#/components/schemas/MarketplaceModel"
          },
          "quantizations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MarketplaceModelQuantization"
            }
          }
        }
      },
...
      "OpenAIError": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "message",
          "type"
        ],
        "properties": {
          "message": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        }
      },
      "OpenAIErrorResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "error",
          "errorCode"
        ],
        "properties": {
          "errorCode": {
            "type": "integer",
            "format": "int32",
            "example": 500
          },
          "error": {
            "$ref": "#/components/schemas/OpenAIError"
          }
        }
      },
...
      "OpenAIModel": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "object",
          "created",
          "owned_by"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "object": {
            "type": "string"
          },
          "created": {
            "type": "integer"
          },
          "owned_by": {
            "type": "string"
          }
        }
      },
      "OpenAIModelList": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "object",
          "data"
        ],
        "properties": {
          "object": {
            "type": "string"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OpenAIModel"
            }
          }
        }
      },
...
      "AnthropicCapabilitySupport": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "supported"
        ],
        "properties": {
          "supported": {
            "type": "boolean"
          }
        }
      },
      "AnthropicContextManagementCapability": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "supported"
        ],
        "properties": {
          "supported": {
            "type": "boolean"
          }
        }
      },
...
      "AnthropicThinkingCapability": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "supported",
          "types"
        ],
        "properties": {
          "supported": {
            "type": "boolean"
          },
          "types": {
            "$ref": "#/components/schemas/AnthropicThinkingTypes"
          }
        }
      },
      "AnthropicModelCapabilities": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "batch",
          "citations",
          "code_execution",
          "context_management",
          "image_input",
          "pdf_input",
          "structured_outputs",
          "thinking"
        ],
        "properties": {
          "batch": {
            "$ref": "#/components/schemas/AnthropicCapabilitySupport"
          },
          "citations": {
            "$ref": "#/components/schemas/AnthropicCapabilitySupport"
          },
          "code_execution": {
            "$ref": "#/components/schemas/AnthropicCapabilitySupport"
          },
          "context_management": {
            "$ref": "#/components/schemas/AnthropicContextManagementCapability"
          },
          "image_input": {
            "$ref": "#/components/schemas/AnthropicCapabilitySupport"
          },
          "pdf_input": {
            "$ref": "#/components/schemas/AnthropicCapabilitySupport"
          },
          "structured_outputs": {
            "$ref": "#/components/schemas/AnthropicCapabilitySupport"
          },
          "thinking": {
            "$ref": "#/components/schemas/AnthropicThinkingCapability"
          }
        }
      },
      "AnthropicModelInfo": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "type",
          "display_name",
          "created_at",
          "max_input_tokens",
          "max_tokens",
          "capabilities"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "display_name": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          },
          "max_input_tokens": {
            "type": "integer"
          },
          "max_tokens": {
            "type": "integer"
          },
          "capabilities": {
            "$ref": "#/components/schemas/AnthropicModelCapabilities"
          }
        }
      },
      "AnthropicModelListResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "data",
          "has_more"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AnthropicModelInfo"
            }
          },
          "first_id": {
            "type": "string"
          },
          "has_more": {
            "type": "boolean"
          },
          "last_id": {
            "type": "string"
          }
        }
      },
      "AnthropicErrorObject": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "type",
          "message"
        ],
        "properties": {
          "type": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "AnthropicErrorResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "type",
          "error",
          "errorCode"
        ],
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "error"
            ]
          },
          "errorCode": {
            "type": "integer",
            "format": "int32",
            "example": 500
          },
          "error": {
            "$ref": "#/components/schemas/AnthropicErrorObject"
          }
        }
      }
    }
  }
}